Fix git 'upgrade-submodules' alias
Fix implementation to work with git 1.9.0. This version of git
introduced changes that were not backward compatible with older versions
of git, breaking the functionality in upgrade-submodules.

Previous versions of git eval'ed the command string passed to `git
submodule foreach`, but newer versions of git pass the command directly
to the shell.