Fix bug with asynchronous prompt
With certain versions of zsh (e.g. 5.0.2 (x86_64-pc-linux-gnu)) and
zsh-syntax-highlighting (e.g. 59fbdda64c21dd9e911329f31fbbedc69123865b,
the earliest commit that triggers this), asynchronously rendering the
initial RPROMPT causes the shell to segfault. This is almost certainly a
bug in zsh (and not zsh-syntax-highlighting or the prompt code).

One way of dealing with the issue is to do the first RPROMPT rendering
in a special way. We could:

* Synchronously render the initial RPROMPT
* Skip rendering the initial RPROMPT

This patch takes the second approach in the interest of minimizing the
time to display the initial prompt.