Drop compatibility with tmux < 2.4
This patch removes all the conditional configuration in tmux.conf.

Note: the old conditional configuration for the color scheme, introduced
in 31170e6fd47d75906a3051d5c407c15d0b4dc0da, was missing the `| vercmp`,
so the buggy conditional always returned false (because of the syntax
error). This ended up working fine in practice, because it turns out
that the conditional was unnecessary as long as the tmux version was
greater than 1.9. The individual `-fg`, `-bg`, and `-attr` options were
superseded by `-style` options back in tmux 1.9, but the deprecated
options were only removed in tmux 2.9.